clang version 21.1.8 clean
modernize-use-integer-sign-comparison
- Added CHECK_EQUAL + REQUIRE_EQUAL for safe signed/unsigned comparison,
  similar to std::cmp_equal
  - used in test_latch01.cpp
- jau::cfmt::impl::append_integral: Fix
- jau::io::ByteStream_File::peek: NOLINT -> false positive

bugprone-bitwise-pointer-cast
- jau::pointer_cast NOLINT -> intentional

bugprone-return-const-ref-from-parameter
- Adding deleted overload for
  - fraction<T> min/max
  - fraction_timespec<T> min/max
  - mp::BigInt min/max

misc-use-internal-linkage
- test_enumutil.cpp: NOLINT -> intentional

+++ include/jau/string_cfmt.hpp
@@ -465,7 +465,7 @@ namespace jau::cfmt {
 
     namespace impl {
         inline constexpr const size_t char32buf_maxlen = 32;
-        inline constexpr const size_t default_float_precision = 6;
+        inline constexpr const uint32_t default_float_precision = 6;
         inline constexpr const double_t max_append_float = (double_t)1e9;
 
         void append_rev(std::string &dest, const size_t dest_maxlen, std::string_view src, bool prec_cut, bool reverse, const FormatOpts &opts);
